Trait-based API
